restructured EFB Apis
