+++ fbw-a32nx/src/systems/instruments/src/EFB/Store/features/simBrief.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { getSimbriefData } from '../../SimbriefApi';
-import { IFuel, IWeights } from '../../SimbriefApi/simbriefInterface';
+import { getSimbriefData } from '../../Apis/Simbrief';
+import { IFuel, IWeights } from '../../Apis/Simbrief/simbriefInterface';
 
 import { store, RootState } from '../store';
