+++ src/instruments/src/EFB/Apis/ChartFox/ChartFox.tsx
