+++ src/instruments/src/EFB/Apis/Navigraph/Navigraph.tsx
